Extract capability probe core into a service
Script becomes a thin CLI over LlmCapabilityProbe so the same
core can run from the dev-area jobs runner.
